requirements.txt
streamlit
langchain 
pydantic 
lxml
langchain-google-genai
google.generativeai
chromadb
google-generativeai
langsmith 
langchainhub
openai 
tiktoken 
opencv-python
tensorflow-probability
imageio 
pillow




